Move shared TablePagination to common components
